Added modules& functionallity for checkbox
