+++ .eslintrc.json
@@ -22,7 +22,8 @@
                 "js": "always",
                 "json": "always"
             }
-        ]
+        ],
+        "import/no-mutable-exports": "off"
     },
     "ignorePatterns": [
         "dist/",
