+++ package.json
@@ -21,7 +21,9 @@
   },
   "homepage": "https://github.com/MohamedRoshdy2021/todo-list#readme",
   "devDependencies": {
+    "@babel/preset-env": "^7.22.9",
     "babel-eslint": "^10.1.0",
+    "babel-loader": "^9.1.3",
     "css-loader": "^6.8.1",
     "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
