+++ src/index.html
@@ -9,9 +9,9 @@
 <body>
   <div id="todo-list">
     <h3>To-Do List</h3>
-    <div class="input-div"><input type="text" id="task-input" placeholder="Add to your list... or double click if you want to edit it!"><i class="fa-solid sign fa-arrow-up-right-from-square"></i></div>
+    <div class="input-div"><input type="text" id="task-input" placeholder="Add to your list.. or double click on text if you want to edit it!"><i class="fa-solid sign fa-arrow-up-right-from-square"></i></div>
     <div id="list-item"></div>
-    <a id="clearall" >clear everything</a>
+    <a id="clearall" >Clear Completed Tasks</a>
   </div>
 </body>
 </html>
