+++ src/style.css
@@ -10,7 +10,7 @@
   display: flex;
   flex-direction: column;
   justify-content: end;
-  box-shadow: 10px 8px 8px rgb(227, 226, 226);
+  box-shadow: 2px 4px 16px 14px rgb(255, 232, 232);
 }
 
 #todo-list h3 {
